+++ src/App.tsx
@@ -2,19 +2,8 @@ import Router from "@/components/Router";
 import { ThemeProvider } from "styled-components";
 import { GlobalStyle } from "@/styles/global-style";
 import { theme } from "@/styles/theme";
-import { RecoilRoot, useRecoilSnapshot } from "recoil";
-import { useEffect } from "react";
-
-const DebugObserver = () => {
-  const snapshot = useRecoilSnapshot();
-  useEffect(() => {
-    console.log("The following atoms were modified:");
-    for (const node of snapshot.getNodes_UNSTABLE({ isModified: true })) {
-      console.log(node.key, snapshot.getLoadable(node));
-    }
-  }, [snapshot]);
-  return null;
-};
+import { RecoilRoot } from "recoil";
+import { DebugObserver } from "@/components/common";
 
 function App() {
   return (
