+++ src/components/common/DebugObserver.tsx
@@ -1,0 +1,15 @@
+import { useEffect } from "react";
+import { useRecoilSnapshot } from "recoil";
+
+const DebugObserver = () => {
+  const snapshot = useRecoilSnapshot();
+  useEffect(() => {
+    console.log("The following atoms were modified:");
+    for (const node of snapshot.getNodes_UNSTABLE({ isModified: true })) {
+      console.log(node.key, snapshot.getLoadable(node));
+    }
+  }, [snapshot]);
+  return null;
+};
+
+export default DebugObserver;
