+++ src/state/user.ts
@@ -2,5 +2,5 @@ import { atom } from "recoil";
 
 export const userState = atom({
   key: "userState",
-  default: { id: 201900810, token: "" },
+  default: { token: "" },
 });
